test_3/test_2/simpledms/requirements.txt
Django==3.2.3
django-filter==2.4.0
djangorestframework==3.12.4
Markdown==3.3.4
openpyxl==3.0.7
drf-yasg
celery
redis
flower
